requirements.txt
langchain
ipykernel
pandas
numpy
scikit_learn
langchain_community
pypdf
openai
langsmith
langchain_openai
python-dotenv
transformers
torch
langchain-huggingface
langgraph
huggingface_hub
groq
langchain-groq
mlflow
sentence_transformers
faiss-cpu
streamlit
qdrant_client
langchain-mcp-adapters
mcp
asyncio
mcp[cli]
ddgs